refactor: DIDEx Rest Notification - Wrapped Connection in RestAPI specific model
- Moved webhook model inside didex package
- Wrapped connection details in webhook specific model
